fix: cloudflare deployment issue
