--- packages/runtime/src/edge.ts
@@ -1,1 +1,0 @@
-index.ts
