+++ packages/runtime/src/enhancements/edge/create-enhancement.ts
@@ -1,0 +1,1 @@
+../node/create-enhancement.ts
