+++ packages/runtime/src/enhancements/edge/default-auth.ts
@@ -1,0 +1,1 @@
+../node/default-auth.ts
