+++ packages/runtime/src/enhancements/edge/delegate.ts
@@ -1,0 +1,1 @@
+../node/delegate.ts
