+++ packages/runtime/src/enhancements/edge/index.ts
@@ -1,0 +1,1 @@
+../node/index.ts
