+++ packages/runtime/src/enhancements/edge/logger.ts
@@ -1,0 +1,1 @@
+../node/logger.ts
