+++ packages/runtime/src/enhancements/edge/omit.ts
@@ -1,0 +1,1 @@
+../node/omit.ts
