+++ packages/runtime/src/enhancements/edge/password.ts
@@ -1,0 +1,1 @@
+../node/password.ts
