+++ packages/runtime/src/enhancements/edge/policy/handler.ts
@@ -1,0 +1,1 @@
+../../node/policy/handler.ts
