+++ packages/runtime/src/enhancements/edge/policy/index.ts
@@ -1,0 +1,1 @@
+../../node/policy/index.ts
