+++ packages/runtime/src/enhancements/edge/promise.ts
@@ -1,0 +1,1 @@
+../node/promise.ts
