+++ packages/runtime/src/enhancements/edge/proxy.ts
@@ -1,0 +1,1 @@
+../node/proxy.ts
