+++ packages/runtime/src/enhancements/edge/query-utils.ts
@@ -1,0 +1,1 @@
+../node/query-utils.ts
