+++ packages/runtime/src/enhancements/edge/types.ts
@@ -1,0 +1,1 @@
+../node/types.ts
