+++ packages/runtime/src/enhancements/edge/utils.ts
@@ -1,0 +1,1 @@
+../node/utils.ts
