+++ packages/runtime/src/enhancements/node/index.ts
@@ -1,4 +1,4 @@
-export * from '../cross';
+export * from '../../cross';
 export * from './create-enhancement';
 export * from './types';
 export * from './utils';
