+++ packages/runtime/src/index.ts
@@ -1,7 +1,7 @@
 export * from './constants';
-export * from './enhancements';
+export * from './cross';
+export * from './enhance';
 export * from './error';
 export * from './types';
 export * from './validation';
 export * from './version';
-export * from './enhance';
