+++ packages/runtime/src/types.ts
@@ -1,5 +1,7 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
+import type { z } from 'zod';
+
 export type PrismaPromise<T> = Promise<T> & Record<string, (args?: any) => PrismaPromise<any>>;
 
 /**
@@ -86,4 +88,81 @@ export type CrudContract = Record<string, DbOperations>;
  */
 export type DbClientContract = CrudContract & {
     $transaction: <T>(action: (tx: CrudContract) => Promise<T>, options?: unknown) => Promise<T>;
+};
+
+/**
+ * Transaction isolation levels: https://www.prisma.io/docs/orm/prisma-client/queries/transactions#transaction-isolation-level
+ */
+export type TransactionIsolationLevel =
+    | 'ReadUncommitted'
+    | 'ReadCommitted'
+    | 'RepeatableRead'
+    | 'Snapshot'
+    | 'Serializable';
+
+/**
+ * Options for enhancing a PrismaClient.
+ */
+export type EnhancementOptions = {
+    /**
+     * The kinds of enhancements to apply. By default all enhancements are applied.
+     */
+    kinds?: EnhancementKind[];
+
+    /**
+     * Whether to log Prisma query
+     */
+    logPrismaQuery?: boolean;
+
+    /**
+     * Hook for transforming errors before they are thrown to the caller.
+     */
+    errorTransformer?: ErrorTransformer;
+
+    /**
+     * The `maxWait` option passed to `prisma.$transaction()` call for transactions initiated by ZenStack.
+     */
+    transactionMaxWait?: number;
+
+    /**
+     * The `timeout` option passed to `prisma.$transaction()` call for transactions initiated by ZenStack.
+     */
+    transactionTimeout?: number;
+
+    /**
+     * The `isolationLevel` option passed to `prisma.$transaction()` call for transactions initiated by ZenStack.
+     */
+    transactionIsolationLevel?: TransactionIsolationLevel;
+};
+
+/**
+ * Context for creating enhanced `PrismaClient`
+ */
+export type EnhancementContext<User extends AuthUser = AuthUser> = {
+    user?: User;
+};
+
+/**
+ * Kinds of enhancements to `PrismaClient`
+ */
+export type EnhancementKind = 'password' | 'omit' | 'policy' | 'validation' | 'delegate';
+
+/**
+ * Function for transforming errors.
+ */
+export type ErrorTransformer = (error: unknown) => unknown;
+
+/**
+ * Zod schemas for validation
+ */
+export type ZodSchemas = {
+    /**
+     * Zod schema for each model
+     */
+    models: Record<string, z.ZodSchema>;
+
+    /**
+     * Zod schema for Prisma input types for each model
+     */
+    input?: Record<string, Record<string, z.ZodSchema>>;
 };
