+++ packages/schema/src/plugins/enhancer/policy/policy-guard-generator.ts
@@ -96,11 +96,14 @@ export class PolicyGenerator {
                 { name: 'type QueryContext' },
                 { name: 'type CrudContract' },
                 { name: 'allFieldsEqual' },
-                { name: 'type PolicyDef' },
                 { name: 'type PermissionCheckerContext' },
-                { name: 'type PermissionCheckerConstraint' },
             ],
             moduleSpecifier: `${RUNTIME_PACKAGE}`,
+        });
+
+        sf.addImportDeclaration({
+            namedImports: [{ name: 'type PolicyDef' }, { name: 'type PermissionCheckerConstraint' }],
+            moduleSpecifier: `${RUNTIME_PACKAGE}/enhancements/node`,
         });
 
         // import enums
